Updated implementation of RotateCopy
A bug in Mesh::locate_triangle emerged during the insert_internal_contours stage of initial mesh refinement for Stator0 use case. This is related to the locate_triangle method supplying incorrect return values for either nonconvex domains, boundary points, or both, when starting the search from particular edges. The implementation of locate_triangle needs to be reconsidered. Perhaps a bounding box needs to be stored along with the mesh of the actual domain.